remove uint32 downcasts

+++ indexfile/mph.go
@@ -235,9 +235,9 @@ type Table struct {
 	mm         *mmap.ReaderAt
 	offsets    []byte
 	level0     []byte
-	level0Mask uint32
+	level0Mask uint64
 	level1     []byte
-	level1Mask uint32
+	level1Mask uint64
 }
 
 // NewTable returns a new `*indexfile.Table` based on the on-disk table at `path`.
@@ -262,16 +262,16 @@ func NewTable(path string) (*Table, error) {
 		return nil, fmt.Errorf("madvise: %s", err)
 	}
 
-	offsetLen := binary.LittleEndian.Uint32(m[8:12])
-	level0Len := binary.LittleEndian.Uint32(m[12:16])
-	level1Len := binary.LittleEndian.Uint32(m[16:20])
+	offsetLen := uint64(binary.LittleEndian.Uint32(m[8:12]))
+	level0Len := uint64(binary.LittleEndian.Uint32(m[12:16]))
+	level1Len := uint64(binary.LittleEndian.Uint32(m[16:20]))
 
 	rest := m[fileHeaderSize:]
 	offsets := rest[:offsetLen*8]
 	level0 := rest[offsetLen*8 : offsetLen*8+level0Len*4]
 	level1 := rest[offsetLen*8+level0Len*4:]
 
-	if uint32(len(level1)) != level1Len*4 {
+	if uint64(len(level1)) != level1Len*4 {
 		return nil, fmt.Errorf("bad len for level1: %d (expected %d)", len(level1), level1Len)
 	}
 
@@ -292,9 +292,9 @@ func (t *Table) MaybeLookupString(s string) uint64 {
 
 // MaybeLookup searches for b in t and returns its potential index.
 func (t *Table) MaybeLookup(b []byte) uint64 {
-	i0 := uint64(uint32(farm.Hash64WithSeed(b, 0)) & t.level0Mask)
+	i0 := farm.Hash64WithSeed(b, 0) & t.level0Mask
 	seed := uint64(binary.LittleEndian.Uint32(t.level0[i0*4 : i0*4+4]))
-	i1 := uint64(uint32(farm.Hash64WithSeed(b, seed)) & t.level1Mask)
+	i1 := farm.Hash64WithSeed(b, seed) & t.level1Mask
 	n := binary.LittleEndian.Uint32(t.level1[i1*4 : i1*4+4])
 	return binary.LittleEndian.Uint64(t.offsets[n*8 : n*8+8])
 }
